Backend/requirements.txt
Django==1.11.28
django-filter==2.1.0
django-jet==1.0.8
djangorestframework==3.9.2
Pillow==5.4.1
pytz==2018.9
